feat: init payment screen
