+++ src/const/theme.const.ts
@@ -41,8 +41,7 @@ export const CUSTOM_THEME = {
 
 const SCROLL_VIEW_DEFAULT: ViewStyle = {
   flex: 1,
-  paddingBottom: 16,
-  maxHeight: Dimensions.get('window').height - CUSTOM_THEME.headerHeight - 16,
+  maxHeight: Dimensions.get('window').height - CUSTOM_THEME.headerHeight
 };
 
 type BASE_STYLE_KEY = 'CONTAINER_WRAP_BOT_BTN' | 'SCROLL_VIEW_DEFAULT' | 'SCROLL_VIEW_BOT_BTN' | 'SURFACE_DEFAULT';
@@ -61,7 +60,8 @@ export const BASE_STYLE: BASE_STYPE_TYPE = {
   SCROLL_VIEW_BOT_BTN: {
     ...SCROLL_VIEW_DEFAULT,
     // def pad-bot: 16, bot btn pat bot 48
-    maxHeight: (SCROLL_VIEW_DEFAULT.maxHeight as number) - 48,
+    paddingBottom: 16,
+    maxHeight: (SCROLL_VIEW_DEFAULT.maxHeight as number) - 48 - 16,
   },
 
   SURFACE_DEFAULT: {
