+++ src/screens/App/AppNavigator.tsx
@@ -1,7 +1,7 @@
 import { createStackNavigator } from '@react-navigation/stack';
 import { AppHeader } from 'components/header';
 import { Screen } from 'types';
-import { CartScreen, CheckoutFinishScreen } from 'screens/Cart';
+import { CartScreen, CheckoutFinishScreen, PaymentScreen } from 'screens/Cart';
 import { AllBranchesScreen, ProductDetailsScreen, ProductListScreen } from 'screens/Common';
 import { HomeScreen } from 'screens/Home';
 
@@ -14,6 +14,7 @@ const AppNavigator = () => {
     <Stack.Navigator screenOptions={{ header: renderAppHeader }}>
       <Stack.Screen name={Screen.Home} component={HomeScreen} />
       <Stack.Screen name={Screen.Cart} component={CartScreen} />
+      <Stack.Screen name={Screen.Payment} component={PaymentScreen} />
       <Stack.Screen name={Screen.CheckoutFinish} component={CheckoutFinishScreen} />
       <Stack.Screen name={Screen.ProductList} component={ProductListScreen} />
       <Stack.Screen name={Screen.ProductDetail} component={ProductDetailsScreen} />
