+++ src/screens/Cart/index.ts
@@ -1,4 +1,5 @@
 import CartScreen from './Cart.screen';
 import CheckoutFinishScreen from './CheckoutFinish.screen';
+import PaymentScreen from './Payment.screen';
 
-export { CartScreen, CheckoutFinishScreen };
+export { CartScreen, CheckoutFinishScreen , PaymentScreen};
